% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{calZscore.default}
\alias{calZscore.default}
\title{Calculate z-score of a vector}
\usage{
\method{calZscore}{default}(x, mad = T)
}
\arguments{
\item{x}{a vector of values or a sample dataset object}

\item{mad}{binary, whether use Median Absolute Deviation to calculate the
standard deviation}
}
\value{
a vector of z-scores
}
\description{
Calculate z-scores of a vector
}
