% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{calZscore.sampleDataset}
\alias{calZscore.sampleDataset}
\title{Calculate z-scores of QC metrics of a SampleDataset}
\usage{
\method{calZscore}{sampleDataset}(object, strat = NULL, qcMetrics = NULL, mad = T, maxVcfZ = F)
}
\arguments{
\item{object}{a sampleDataset object.}

\item{strat}{optional, logical, variable, or vector of attributes the sample
will be stratified by.}

\item{qcMetrics}{optional, a vector of names of QC metrics to compute z-score.}

\item{mad}{optional, binary, whether to use Median Absolute Deviation to
calculate z-score or not, default is TRUE.}

\item{maxVcfZ}{optional, logical, calculate to report maximum Vcf Z-score.}
}
\description{
Calculate z-scores of QC metrics of a SampleDataset
}
