% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{dim.sampleDataset}
\alias{dim.sampleDataset}
\title{Show dimensions of sample dataset object}
\usage{
\method{dim}{sampleDataset}(object)
}
\arguments{
\item{object}{sample data set}
}
\value{
a vector of rows and columns of the data frame of sample data set
}
\description{
Show dimensions of sample dataset object
}
