% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{.calMaxVcfZ}
\alias{.calMaxVcfZ}
\title{Calculate maximum vcf z-score and update the object with a maxVcfzscore
column}
\usage{
.calMaxVcfZ(object, qcMetrics)
}
\description{
Calculate maximum vcf z-score and update the object with a maxVcfzscore
column
}
