% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.multiplotWithSharedLegend}
\alias{.multiplotWithSharedLegend}
\title{Generate multiple plots with shared figure legend}
\usage{
.multiplotWithSharedLegend(plots, ncols = 5, position = c("bottom",
  "right"), show = T)
}
\description{
:param plots a list of grob objects
:param nrows number of rows
}
\details{
:Returns a grid graphical object (grob)
}
