% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.scatter}
\alias{.scatter}
\title{Produce a scatter plot from a data frame}
\usage{
.scatter(data, x, y, strat, xlab = NULL, ylab = NULL, outliers = NULL,
  main = NULL, legend = T, primaryID = NULL)
}
\description{
param data a data.frame
param strat by which column to stratify
param outliers a vector of outliers
}
