% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSamples.R
\name{flagSamples}
\alias{flagSamples}
\title{Flag a sample}
\usage{
flagSamples(...)
}
\description{
Flag a sample from a data frame or a sampleDataset
}
\seealso{
{\link{flagSamples.default}}

{\link{flagSamples.sampleDataset}}
}
