% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSamples.R
\name{flagSamples.sampleDataset}
\alias{flagSamples.sampleDataset}
\title{Flag a sample from a sample data set}
\usage{
\method{flagSamples}{sampleDataset}(object, cutoffs = NULL, zscore = NULL)
}
\arguments{
\item{object}{sample data set}

\item{cutoffs}{a data frame}

\item{zscore}{zscore cutoff}
}
\value{
an updated sample dataset object with flags added.
}
\description{
Flag a sample from a sample data set
}
