% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Create a split violin plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data for this layer}

\item{position}{Position adjustment}

\item{...}{Other arguments passed to the layer}

\item{draw_quantiles}{Quantiles to be drawn}

\item{trim}{Should the data be trimmed?}

\item{scale}{Area or count for width of violin}

\item{na.rm}{Should missing values be removed?}

\item{show.legend}{Should this layer be included in the legends?}

\item{inherit.aes}{Should inherit aesthetics from the parent plot?}
}
\value{
A ggplot2 layer for a split violin plot
}
\description{
Create a split violin plot
}
\seealso{
{\link{outlierPlots}}

{\link{outlierPlots.default}}

{\link{outlierPlots.sampleDataset}}

{\link{PCplots}}

{\link{multiplot}}

{\link{scatter}}

{\link{multiplotWithSharedLegend}}
}
