% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttr.R
\name{getAttr}
\alias{getAttr}
\title{Get attributes of an sds}
\usage{
getAttr(sds, ...)
}
\arguments{
\item{sds}{sample dataset}

\item{attribute}{name of attribute to return}

\item{showID}{whether to display sample IDs or not in the output data frame}
}
\value{
a data frame that contains values of specific attribute
}
\description{
Get attributes of an sds
}
\seealso{
{\link{getAttr.sampleDataset}}
}
