% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttr.R
\name{getAttr.sampleDataset}
\alias{getAttr.sampleDataset}
\title{Get attributes of a SampleDataset}
\usage{
\method{getAttr}{sampleDataset}(sds, attribute, showID = F)
}
\arguments{
\item{sds}{sample data set}

\item{attribute}{name of attribute to return}

\item{showID}{whether to display sample IDs or not in the output data frame}
}
\value{
a data frame that contains values of specific attribute
}
\description{
Get attributes of a SampleDataset
}
