% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasAttr.R
\name{hasAttr}
\alias{hasAttr}
\title{Check whether the sampleDataset has attributes provided}
\usage{
hasAttr(object, attr)
}
\arguments{
\item{object}{a sampleDataset object}

\item{attr}{attributes}
}
\value{
a boolean value
}
\description{
Check whether the sampleDataset has attributes provided
}
