% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry.default}
\alias{inferAncestry.default}
\title{Infer anestry with KNN}
\usage{
\method{inferAncestry}{default}(testSet, trainSet, ancestry, k = 5)
}
\arguments{
\item{testSet}{matrix used for test}

\item{trainSet}{matrix for training PCs}

\item{ancestry}{ancestries correspond to the training PCs}

\item{k}{Number of nearest neighbour to use in this classfication}
}
\description{
Infer anestry with KNN
}
