% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry}
\alias{inferAncestry}
\title{Predict sample ancestry from genotype Principle Components}
\usage{
inferAncestry(...)
}
\value{
Updated sample Dataset object with inferredAncestry tag
}
\description{
Principle Components of genotypes with k nearest neighbours.
}
