% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry.sampleDataset}
\alias{inferAncestry.sampleDataset}
\title{Infer sample ancestry}
\usage{
\method{inferAncestry}{sampleDataset}(sds, trainSet, knownAncestry, nPC = 3, k = 5)
}
\arguments{
\item{sds}{Sample Dataset sds for Ancestry predition}

\item{trainSet}{data frame with genotype PCs for each sample with known ancestry}

\item{knownAncestry}{a vector with known ancestry for each sample in same order of trainSet}

\item{nPC}{Use first n PCs to predict ancestry}
}
\description{
Infer sample ancestry
}
