% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Generate multi-panel plots}
\usage{
multiplot(..., plotlist = NULL, file = NULL, ncols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot2 objects}

\item{plotlist}{a list of ggplot2 objects}

\item{file}{output file name (optional)}

\item{ncols}{number of columns in the plot}

\item{layout}{matrix specifying the layout of panels}
}
\value{
The multi-panel plot
}
\description{
This function creates a multi-panel plot from a list of ggplot2 objects.
}
\examples{
\dontrun{
# Example usage:
multiplot(plot1, plot2, ncols = 2)
}

}
\seealso{
{\link{outlierPlots}}

{\link{outlierPlots.default}}

{\link{outlierPlots.sampleDataset}}

{\link{PCplots}}

{\link{scatter}}

{\link{multiplotWithSharedLegend}}

{\link{geom_split_violin}}
}
