% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplotWithSharedLegend}
\alias{multiplotWithSharedLegend}
\title{Generate multiple plots with shared figure legend}
\usage{
multiplotWithSharedLegend(
  plots,
  ncols = 5,
  position = c("bottom", "right"),
  show = T
)
}
\arguments{
\item{plots}{a list of grob sdss}

\item{ncols}{number of rows}

\item{position}{Position of the legend, between "bottom" and "right"}

\item{show}{whether show the plot or not}
}
\value{
a grid graphical `sds` (grob)
}
\description{
Generate multiple plots with shared figure legend
}
\seealso{
{\link{outlierPlots}}

{\link{outlierPlots.default}}

{\link{outlierPlots.sampleDataset}}

{\link{PCplots}}

{\link{multiplot}}

{\link{scatter}}

{\link{geom_split_violin}}
}
