% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{outlierPlots.default}
\alias{outlierPlots.default}
\title{Generate outlier plots for a dataset}
\usage{
\method{outlierPlots}{default}(
  tab,
  qcMetrics,
  strat,
  main,
  outliers,
  primaryID = NULL,
  type = "violin"
)
}
\arguments{
\item{tab}{Input table with data.}

\item{qcMetrics}{Metrics to plot.}

\item{strat}{Factor by which to stratify the dots.}

\item{main}{Title of the plot.}

\item{outliers}{List of outliers.}

\item{primaryID}{Identifier for samples (if available).}

\item{type}{Type of plot ('violin' or 'density').}
}
\value{
A list of outlier plots.
}
\description{
Generate outlier plots for a dataset
}
\examples{
\dontrun{
# Example usage:
outlierPlots.default(tab, qcMetrics = "Metric1", strat = "Group", main = "Outlier Plot")
}

}
\seealso{
{\link{outlierPlots}}

{\link{outlierPlots.sampleDataset}}

{\link{PCplots}}

{\link{multiplot}}

{\link{scatter}}

{\link{multiplotWithSharedLegend}}

{\link{geom_split_violin}}
}
