% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{outlierPlots}
\alias{outlierPlots}
\title{Plot outliers}
\usage{
outlierPlots(tab, qcMetrics, strat, main, outliers, type = "violin")
}
\arguments{
\item{tab}{input table}

\item{qcMetrics}{which QC metrics}
}
\description{
Plot outliers
}
