% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{outlierPlots.sampleDataset}
\alias{outlierPlots.sampleDataset}
\title{Produce outlier plots for a sampleDataset}
\usage{
\method{outlierPlots}{sampleDataset}(sds, title, width = 15, height = 6)
}
\value{
A pdf with all plots
}
\description{
Produce outlier plots for a sampleDataset
}
