% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{print.sampleDataset}
\alias{print.sampleDataset}
\title{Print information about the sample dataset `sds`.}
\usage{
\method{print}{sampleDataset}(sds)
}
\arguments{
\item{sds}{Sample dataset `sds`.}
}
\description{
Print information about the sample dataset `sds`.
}
