% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDataset.R
\name{sampleDataset}
\alias{sampleDataset}
\title{Sample dataset object}
\usage{
sampleDataset(bamQcMetr, vcfQcMetr, attributes, primaryID = NULL, df = NULL)
}
\arguments{
\item{bamQcMetr}{if df is provided, this is a vector of names}

\item{vcfQcMetr}{a data frame contain}

\item{attributes}{}

\item{df}{an R data.frame that contains all sample level data, including QC metrics and}
}
\value{
an sampleDataset object
}
\description{
Create Object sample dataset object to explore outliers.
}
