% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{save}
\alias{save}
\title{Save an sds to a specific format}
\usage{
save(sds, ...)
}
\arguments{
\item{sds}{sample dataset}

\item{...}{Additional arguments to be passed.}
}
\description{
Save an sds to tsv, RDS or excel format
}
\seealso{
{\link{sort.sampleDataset}}

{\link{save.sampleDataset}}

{\link{dim.sampleDataset}}

{\link{print.sampleDataset}}
}
