% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{save.sampleDataset}
\alias{save.sampleDataset}
\title{Write sample dataset to tsv, RDS or excel files with S3 method save.}
\usage{
\method{save}{sampleDataset}(object, RDS = NULL, tsv = NULL, xls = NULL)
}
\arguments{
\item{object}{sample dataset}

\item{RDS}{path and output name of RDS file}

\item{tsv}{path and output name of tsv file}

\item{xls}{path and output name of excel file}
}
\description{
Write sample dataset to tsv, RDS or excel files with S3 method save.
}
