% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatter}
\alias{scatter}
\title{Produce a scatter plot from a data frame}
\usage{
scatter(
  data,
  x,
  y,
  strat,
  xlab = NULL,
  ylab = NULL,
  outliers = NULL,
  main = NULL,
  legend = T,
  primaryID = NULL
)
}
\arguments{
\item{data}{a data.frame}

\item{strat}{by which column to stratify}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{outliers}{a vector of outliers}

\item{main}{main title of the plot}

\item{legend}{binary, whether to plot legend or not}

\item{primaryID}{the primary ID used to}
}
\description{
Produce a scatter plot from a data frame
}
