% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr}
\alias{setAttr}
\title{Update a Sample Data Set}
\usage{
setAttr(object, ...)
}
\arguments{
\item{object}{sample dataset object}

\item{...}{other arguments}
}
\value{
object
}
\description{
Update entries of sample data set, such as the df, bamQcMetr, vcfQcMetr
}
