% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr.sampleDataset}
\alias{setAttr.sampleDataset}
\title{Set attributes of a sample dataset}
\usage{
\method{setAttr}{sampleDataset}(object, attributes, data, primaryID,
  overwrite = F)
}
\arguments{
\item{object}{sampleDataset}

\item{entry}{which entry to add to sample Dataset}

\item{dataframe}{content to add to the dataframe}

\item{by}{primary key of input data frame}
}
\value{
updated sampleDataset object
}
\description{
Set attributes of a sample dataset
}
