% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{sort.sampleDataset}
\alias{sort.sampleDataset}
\title{Update index of sample dataset by different annotation categories}
\usage{
\method{sort}{sampleDataset}(sds, by)
}
\arguments{
\item{sds}{sample dataset}

\item{by}{sort by which category}
}
\value{
an sample dataset sds with updated index
}
\description{
Update index of sample dataset by different annotation categories
}
