% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{calZscore}
\alias{calZscore}
\title{Calculate z-score of a vector}
\usage{
calZscore(x, ...)
}
\arguments{
\item{x}{a vector of values.}

\item{mad}{binary, whether use median absolute deviation, if false, standard deviation will be used.}
}
\value{
a vector of z-scores
}
\description{
Calculate z-score of a vector
}
