% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{calZscore}
\alias{calZscore}
\title{Calculate z-score of a vector or QC metrics of a SampleDataset}
\usage{
calZscore(...)
}
\value{
The form of the value returned by predict depends on the class of
        its argument. See the documentation of the particular methods for
        details of what is produced by that method.
}
\description{
Calculate z-score of a vector or QC metrics of a SampleDataset,
             see details in calZscore.default and calZscore.sampleDataset
}
