% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{calZscore.sampleDataset}
\alias{calZscore.sampleDataset}
\title{Calculate z-scores of QC metrics of a SampleDataset}
\usage{
\method{calZscore}{sampleDataset}(object, strat = NULL, qcMetrics = NULL, mad = T, maxVcfZ = F)
}
\arguments{
\item{object}{a sampleDataset object}

\item{strat}{optional, logical, variable or vector of attributes the sample
will be stratified by}

\item{qcMetrics}{optional, a vector of names of qc metrics to compute zscore}

\item{mad}{optional, binary, whether use Median Absolute Deviation to
calculate z-score or not, default is True}

\item{maxVcfZ}{optional, logical, calculate to report maximun Vcf Zscore}
}
\value{
sampleDataset object, with additional attributes: zscoreBy, zscore
        and maximumVcfZscore
}
\description{
Calculate z-scores of QC metrics of a SampleDataset
}
