% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{calZscore.sampleDataset}
\alias{calZscore.sampleDataset}
\title{Calculate z-score of sample dataset}
\usage{
\method{calZscore}{sampleDataset}(object, strat, qcMetrics = NULL)
}
\arguments{
\item{object}{a sample data set}

\item{strat}{variable or vector of attributes the sample will be stratified by}

\item{qcMetrics}{optional, which QC metrics should be stratified on.}
}
\value{
sampleDataset object
}
\description{
Calculate z-score of sample dataset
}
