% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{dim.sampleDataset}
\alias{dim.sampleDataset}
\title{Show dimensions of sample dataset sds}
\usage{
\method{dim}{sampleDataset}(sds)
}
\arguments{
\item{sds}{sample data set sds}
}
\value{
a vector of rows and columns of the data frame of sample data set
}
\description{
Show dimensions of sample dataset sds
}
