% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSamples.R
\name{flagSamples.default}
\alias{flagSamples.default}
\title{Flag a sample from a data frame}
\usage{
\method{flagSamples}{default}(df, stat, cutoff, greater = TRUE, abs = F)
}
\arguments{
\item{df}{a data frame contained summary statistics}
}
\value{
Return a data frame
}
\description{
Flag a sample from a data frame
}
