% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry.default}
\alias{inferAncestry.default}
\title{Infer ancestry with KNN}
\usage{
\method{inferAncestry}{default}(testSet, trainSet, ancestry, k = 5)
}
\arguments{
\item{testSet}{Matrix used for test.}

\item{trainSet}{Matrix for training PCs.}

\item{ancestry}{Ancestries corresponding to the training PCs.}

\item{k}{Number of nearest neighbors to use in this classification.}
}
\value{
A vector of inferred ancestries.
}
\description{
Infer ancestry with KNN
}
\seealso{
{\link{inferAncestry.sampleDataset}}

{\link{inferAncestry}}
}
