% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry}
\alias{inferAncestry}
\title{Predict sample ancestry from genotype Principle Components}
\usage{
inferAncestry(...)
}
\arguments{
\item{...}{Additional parameters for future expansion.}
}
\value{
Updated sample Dataset sds with inferredAncestry tag.
}
\description{
Principle Components of genotypes with k nearest neighbours.
}
\seealso{
{\link{inferAncestry.sampleDataset}}

{\link{inferAncestry.default}}
}
