% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry}
\alias{inferAncestry}
\title{Predict ancestry of samples from genotype Principle Components}
\usage{
inferAncestry(object, ...)
}
\arguments{
\item{object}{Sample Dataset object for ancestry predition}

\item{trainSet}{optional, a data frame with sample Id and ethnicity}

\item{nPC}{Number of genotype PCs used}
}
\value{
updated sample Dataset object with inferredancestry tag
}
\description{
Principle Components of genotypes
}
