% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAncestry.R
\name{inferAncestry.sampleDataset}
\alias{inferAncestry.sampleDataset}
\title{Infer sample ancestry}
\usage{
\method{inferAncestry}{sampleDataset}(sds, trainSet, knownAncestry, nPC = 3, k = 5)
}
\arguments{
\item{sds}{Sample Dataset sds for ancestry prediction.}

\item{trainSet}{Data frame with genotype PCs for each sample with known ancestry.}

\item{knownAncestry}{A vector with known ancestry for each sample in the same order as trainSet.}

\item{nPC}{Use the first n PCs to predict ancestry.}
}
\value{
Updated sample Dataset sds with inferredAncestry tag.
}
\description{
Infer sample ancestry
}
\seealso{
{\link{inferAncestry}}

{\link{inferAncestry.default}}
}
