% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{generate multi panel plots}
\usage{
multiplot(..., plotlist = NULL, file = NULL, ncols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{a list of plot}

\item{file}{file}

\item{ncols}{number of columns in the plot}
}
\description{
generate multi panel plots
}
