% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{outlierPlots.default}
\alias{outlierPlots.default}
\title{Produce outlier plot}
\usage{
\method{outlierPlots}{default}(
  tab,
  qcMetrics,
  strat,
  main,
  outliers,
  primaryID = NULL,
  type = "violin"
)
}
\arguments{
\item{tab}{input table with sample information and metrics}

\item{qcMetrics}{which metrics to plot}

\item{strat}{by which factor to stratify the dots}

\item{main}{title of the plot}

\item{outliers}{vector of sample IDs to highlight as outliers}

\item{primaryID}{the primary ID for labeling}

\item{type}{type of plot, either 'violin' or 'density'}
}
\value{
a list of outlier plots
}
\description{
Produce outlier plot
}
\examples{
outlierPlots.default(tab, qcMetrics = c("Metric1", "Metric2"), strat = "Group",
                      main = "Outlier Plots", outliers = c("SampleA", "SampleB"),
                      primaryID = "SampleID", type = "violin")

}
