% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{outlierPlots}
\alias{outlierPlots}
\title{Produce outlier plot}
\usage{
outlierPlots(...)
}
\arguments{
\item{tab}{Input table with data.}

\item{qcMetrics}{Metrics to plot.}

\item{strat}{Factor by which to stratify the dots.}

\item{main}{Title of the plot.}

\item{outliers}{List of outliers.}

\item{primaryID}{Identifier for samples (if available).}

\item{type}{Type of plot ('violin' or 'density').}
}
\description{
Produce outlier plot
}
\seealso{
{\link{outlierPlots.default}}

{\link{outlierPlots.sampleDataset}}

{\link{PCplots}}

{\link{multiplot}}

{\link{scatter}}

{\link{multiplotWithSharedLegend}}

{\link{geom_split_violin}}
}
