% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{outlierPlots.sampleDataset}
\alias{outlierPlots.sampleDataset}
\title{Produce outlier plots for a sampleDataset}
\usage{
\method{outlierPlots}{sampleDataset}(sds, title, width = 15, height = 6)
}
\arguments{
\item{sds}{A sample dataset object.}

\item{title}{Title for the PDF file.}

\item{width}{Width of the PDF.}

\item{height}{Height of the PDF.}
}
\value{
A PDF file containing outlier plots.
}
\description{
Produce outlier plots for a sampleDataset
}
\examples{
\dontrun{
# Example usage:
outlierPlots.sampleDataset(sds, title = "Outlier_Plots", width = 15, height = 6)
}

}
\seealso{
{\link{outlierPlots.default}}

{\link{outlierPlots}}

{\link{PCplots}}

{\link{multiplot}}

{\link{scatter}}

{\link{multiplotWithSharedLegend}}

{\link{geom_split_violin}}
}
