% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PCplots}
\alias{PCplots}
\title{Generate PC plot of an SampleDataset sds}
\usage{
PCplots(sds, showPlot = T, cor = F, outliers = NULL)
}
\arguments{
\item{sds}{sample data set}

\item{showPlot}{whether to show plot}
}
\value{
a list of ggplot sdss
}
\description{
Generate PC plot of an SampleDataset sds
}
