% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot.default}
\alias{sampleQcPlot.default}
\title{Generate scatter plot of QC metrics}
\usage{
\method{sampleQcPlot}{default}(data, primaryID, qcMetric, annotation = NULL,
  geom = c("scatter", "violin", "hist"), outliers = NULL, legend = T,
  main = "QC")
}
\arguments{
\item{annotation}{which annotation to visualize}

\item{object}{sample dataset object}

\item{pca}{whether perform pca plot}
}
\description{
Generate scatter plot of QC metrics
}
