% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot.default}
\alias{sampleQcPlot.default}
\title{Generate scatter plot of QC metrics}
\usage{
\method{sampleQcPlot}{default}(
  data,
  primaryID,
  qcMetric,
  annot = NULL,
  outliers = NULL,
  legend = T,
  main = "QC",
  geom = c("scatter", "violin", "hist")
)
}
\arguments{
\item{annot}{which annot to visualize}

\item{sds}{sample dataset sds}

\item{pca}{whether perform pca plot}
}
\description{
Generate scatter plot of QC metrics
}
