% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot}
\alias{sampleQcPlot}
\title{Generate scatter plot of QC metrics according to samples}
\usage{
sampleQcPlot(sds, ...)
}
\arguments{
\item{sds}{sample dataset `sds`}

\item{...}{other arguments}
}
\description{
Generate scatter plot of QC metrics according to samples
}
\seealso{
{\link{sampleQcPlot.default}}

{\link{sampleQcPlot.sampleDataset}}
}
