% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot}
\alias{sampleQcPlot}
\title{Generate scatter plot of qcMetrics according to samples}
\usage{
sampleQcPlot(object, ...)
}
\arguments{
\item{object}{sample dataset object}
}
\description{
Generate scatter plot of qcMetrics according to samples
}
