% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot}
\alias{sampleQcPlot}
\title{generate scatter plot of qcMetrics according to samples}
\usage{
sampleQcPlot(...)
}
\arguments{
\item{sds}{sample dataset object}
}
\description{
generate scatter plot of qcMetrics according to samples
}
