% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot.sampleDataset}
\alias{sampleQcPlot.sampleDataset}
\title{Generate scatter plot of QC metrics}
\usage{
\method{sampleQcPlot}{sampleDataset}(object, annotation = NULL, qcMetrics,
  pca = F, inferredAncestry = T, geom = "scatter", outliers = NULL,
  legend = T, main = "QC")
}
\arguments{
\item{object}{sample dataset object}

\item{annotation}{which annotation to visualize}

\item{pca}{whether perform pca plot}

\item{geom}{}
}
\description{
Generate scatter plot of QC metrics
}
