% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleQcPlot.sampleDataset}
\alias{sampleQcPlot.sampleDataset}
\title{Generate a panel of sample QC plots across a list of qcMetrics}
\usage{
\method{sampleQcPlot}{sampleDataset}(
  sds,
  annot,
  qcMetrics = sds$qcMetrics,
  geom = c("scatter", "violin", "hist"),
  outliers = NULL,
  legend = T,
  main = "QC",
  position = c("right", "bottom"),
  ncols = 5,
  show = FALSE,
  sort = TRUE
)
}
\arguments{
\item{sds}{SampleDataset `sds`}

\item{annot}{optional, a character string of sample annot to stratify by}

\item{qcMetrics}{a character string or a vector includes QC metrics to
explore; if unspecified, all QC metrics in the SDS will
be used.}

\item{geom}{a character string indicating which visualization pattern to be
used. One of strings 'scatter', 'violin' or 'hist' can be used.}

\item{outliers}{a vector of IDs of outliers to show on the figure}

\item{legend}{whether to include a legend or not}

\item{position}{position of the legend}

\item{sort}{whether to sort by annot when ploting}
}
\value{
a list of grob sds
}
\description{
Generate a panel of sample QC plots across a list of qcMetrics
}
\seealso{
{\link{sampleQcPlot}}

{\link{sampleQcPlot.sampleDataset}}
}
