% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr}
\alias{setAttr}
\title{Update a Sample Dataset}
\usage{
setAttr(sds, ...)
}
\arguments{
\item{sds}{A sample dataset object}

\item{attributes}{character vector specifying the attributes to add or update in the sample Dataset}

\item{data}{data.frame, content to add to the dataframe}

\item{primaryID}{primary key of input data frame}

\item{overwrite}{logical, indicating whether to overwrite existing attributes if they already exist}
}
\value{
An updated sample dataset
}
\description{
Update entries of a sample dataset, such as the df, bamQcMetr, vcfQcMetr.
}
\seealso{
{\link{setAttr.sampleDataset}}
}
