% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr}
\alias{setAttr}
\title{Update a Sample Data Set}
\usage{
setAttr(sds, ...)
}
\arguments{
\item{sds}{sample dataset sds}

\item{...}{other arguments}
}
\value{
sds
}
\description{
Update entries of sample data set, such as the df, bamQcMetr, vcfQcMetr
}
