% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr}
\alias{setAttr}
\title{Set attributes of a sample dataset}
\usage{
setAttr(object, ...)
}
\arguments{
\item{object}{sample dataset object}

\item{...}{other argumentsoxyg}
}
\value{
object
}
\description{
Set attributes of a sample dataset
}
