% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr.sampleDataset}
\alias{setAttr.sampleDataset}
\title{Set attributes of a sample dataset}
\usage{
\method{setAttr}{sampleDataset}(sds, attributes, data, primaryID, overwrite = F)
}
\arguments{
\item{sds}{sampleDataset}

\item{attributes}{character vector specifying the attributes to add or update in the sample Dataset}

\item{data}{data.frame, content to add to the dataframe}

\item{primaryID}{primary key of input data frame}

\item{overwrite}{logical, indicating whether to overwrite existing attributes if they already exist}
}
\value{
updated sampleDataset sds
}
\description{
Set attributes of a sample dataset
}
