% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttr.R
\name{setAttr.sampleDataset}
\alias{setAttr.sampleDataset}
\title{Set attributes of a sample dataset}
\usage{
\method{setAttr}{sampleDataset}(sds, attributes, data, primaryID, overwrite = F)
}
\arguments{
\item{sds}{sampleDataset}

\item{attributes}{which entry to add to sample Dataset}

\item{data}{data.frame, content to add to the dataframe}

\item{by}{primary key of input data frame}
}
\value{
updated sampleDataset sds
}
\description{
Set attributes of a sample dataset
}
